Added support for joining tables to YDDatabaseTree.
This commit also adds a standard table prefix to the table in the database tree called "t1" which should always be used as a reference in the field definitions.

It was particularly hard to get everything to work as I had to remove/add the table prefix in different parts of the code. Hopefully, I spotted all of them...